// frontend-doctor-app/services/client/types.gen.ts
export type PatientType = 'private_patient' | 'migrant_worker';
// ...
export type SGiMedICType = 'PINK IC' | 'BLUE IC/ENTRY PERMIT' | 'FIN NUMBER' | 'PASSPORT';
// ...
export type TeleconsultResponse = {
    id: string;
    patient_type: PatientType;
    status: TeleconsultStatus;
    doctor_id?: string | null;
    additional_status?: TeleconsultStatus | null;
    queue_number?: string | null;
    checkin_time: string;
    teleconsult_start_time?: string | null;
    branch_name: string;
    user: UserResponse;
};

export type TeleconsultStatus = 'Prepayment' | 'Checked In' | 'Consult Start' | 'Consult End' | 'Outstanding' | 'Dispense Medication' | 'Checked Out' | 'Cancelled' | 'Missed';
// ...
export type UserResponse = {
    name: string;
    ic_type: SGiMedICType;
    nric: string;
};